chore: Refactor kamelets-catalog export
