+++ packages/ui/src/models/index.ts
@@ -2,3 +2,4 @@ export * from './camel-catalog-index';
 export * from './camel-components-catalog';
 export * from './camel-processors-catalog';
 export * from './catalog-kind';
+export * from './kamelets-catalog';
